requirements.txt
pymysql==1.1.1
holidays==0.63
pandas==2.2.3
numpy
pandas-ta==0.3.14b
sqlalchemy<2.0
scikit-learn==1.6.1
pyyaml==6.0.2
setuptools==75.8.0
xgboost==2.1.3
pendulum==3.0.0
settrade-v2
yfinance==0.2.52